Address Coverity new defects
Nine new findings reported by Coverity Scan against current master:

- CID 560615 / 560614 / 560608 (UNINIT) - atomic_int values written
  without atomic_init() in test/threading.c.  C11 requires atomic_init
  for the first touch of a non-statically-initialised atomic, so use
  it instead of plain assignment.

- CID 560613 / 560607 (CHECKED_RETURN) - pthread_mutex_unlock() return
  value ignored in linux_kqueue_start_thread() and
  monitoring_thread_loop().  Cast to (void) to match the dominant call
  pattern (Coverity reports 20/23 sites already do).

- CID 560612 (UNREACHABLE) - dead _exit(0) after for(;;) pause() in
  test_kevent_threading_proc_single_delivery.  Drop it; the comment
  on it already said "unreachable".

- CID 560611 / 560610 / 560609 (INTEGER_OVERFLOW) - the EPOLL_EV_*
  macros assigned the int-typed _events expression to struct
  epoll_event.events (uint32_t) without an explicit cast.  When the
  expression includes EPOLLET (0x80000000) the value is negative under
  the C signed-int promotion rules, and Coverity flags the negative-
  to-unsigned conversion.  Add explicit (uint32_t) casts.  Same
  treatment applied to the two sibling macros (EPOLL_EV_EVENTFD,
  EPOLL_EV) which weren't flagged this round but have the same shape.

Verified: full suite passes on Linux libkqueue (-n 1, 80 tests).

+++ src/linux/platform.h
@@ -232,7 +232,7 @@ void                epoll_udata_defer_free(struct kqueue *kq, struct epoll_udata
  *                           determine the correct filter to notify
  *                           if the event fires.
  */
-#define EPOLL_EV_KN(_events, _kn) &(struct epoll_event){ .events = _events, .data = { .ptr = (_kn)->kn_udata }}
+#define EPOLL_EV_KN(_events, _kn) &(struct epoll_event){ .events = (uint32_t) (_events), .data = { .ptr = (_kn)->kn_udata }}
 
 /** Macro for building a temporary fds epoll_event
  *
@@ -241,20 +241,20 @@ void                epoll_udata_defer_free(struct kqueue *kq, struct epoll_udata
  *                           event. Is used during event demuxing to inform
  *                           multiple filters of read/write events on the fd.
  */
-#define EPOLL_EV_FDS(_events, _fds) &(struct epoll_event){ .events = _events, .data = { .ptr = (_fds)->fds_udata }}
+#define EPOLL_EV_FDS(_events, _fds) &(struct epoll_event){ .events = (uint32_t) (_events), .data = { .ptr = (_fds)->fds_udata }}
 
 /** Macro for building a temporary eventfd epoll_event
  *
  * @param[in] _events        One or more event flags EPOLLIN, EPOLLOUT etc...
  * @param[in] _efd           eventfd to build event for.
  */
-#define EPOLL_EV_EVENTFD(_events, _efd) &(struct epoll_event){ .events = _events, .data = { .ptr = (_efd)->efd_udata }}
+#define EPOLL_EV_EVENTFD(_events, _efd) &(struct epoll_event){ .events = (uint32_t) (_events), .data = { .ptr = (_efd)->efd_udata }}
 
 /** Macro for building a temporary epoll_event
  *
  * @param[in] _events        One or more event flags EPOLLIN, EPOLLOUT etc...
  */
-#define EPOLL_EV(_events) &(struct epoll_event){ .events = _events }
+#define EPOLL_EV(_events) &(struct epoll_event){ .events = (uint32_t) (_events) }
 
 /** Common structure passed as the epoll data.ptr
  *
